[REFACTOR] Controller 기능 분할
Controller에서 View로 입력받아 저장하지 않고
Main에서 View로 입력받은 정보를 Controller에게 전달한다.

Squashed commit of the following:

commit 88386bcc8bd27e5cc977a9e7192a47d6b82ee003
Author: conatuseus <conatuseus@gmail.com>
Date:   Thu May 9 15:28:01 2019 +0900

    Fix code for Refactoring

commit 369b617c9991b0b80f8df7a8082245760e5a9666
Author: conatuseus <conatuseus@gmail.com>
Date:   Thu May 9 15:12:29 2019 +0900

    Fix Code for MVC, Create Main

commit 1afd7d3719978b146b9b7e7920219d37c74e90eb
Author: conatuseus <conatuseus@gmail.com>
Date:   Thu May 9 14:50:07 2019 +0900

    Fix README
