+++ README.md
@@ -33,9 +33,9 @@
 
 [TEST목록]
 - 자동차 이름이 올바르게 입력됐는지 확인(영문자,길이 제한, 중복)
+- 시도할 횟수의 유효성 검사 (1 이상의 정수만 가능하도록)
 - 자동차 리스트에 자동차의 추가가 잘 되었는지 확인
 - 자동차의 이동이 가능할 때 Position의 증가 확인 (이동 불가시 Position 변동 없음 확인)
-- 시도할 횟수의 유효성 검사 (1 이상의 정수만 가능하도록)
 - 4이상의 값이 입력됐을 때(true일 때) position++가 됐는지 확인
 - CarList생성 후 특정 Car를 move(true)하여 maxPosition이  의도한바와 맞는지 확인
