+++ src/main/java/racingcar/RacingCar.java
@@ -1,0 +1,11 @@
+package racingcar;
+
+import racingcar.controller.Controller;
+import racingcar.view.View;
+
+public class RacingCar {
+    public static void main(String[] args) {
+        Controller controller = new Controller();
+        controller.run(View.inputNameOfCar(), View.inputNumberOfTry());
+    }
+}
